pyproject.toml
[tool.poetry]
package-mode = false
name = "teso"
version = "0.1.0"
authors = ["EV9H <github.com/ev9h>"]
description = "Telegram scraping bot using Telethon"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.34.0"
python-dotenv = "^1.0.0"
aiofiles = "^24.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"